Move logic functions from math to logic
