+++ chainerx_cc/chainerx/cuda/cuda_device/misc.cu
@@ -16,6 +16,7 @@
 #include "chainerx/kernels/arithmetic.h"
 #include "chainerx/kernels/math.h"
 #include "chainerx/kernels/misc.h"
+#include "chainerx/kernels/logic.h"
 #include "chainerx/routines/type_util.h"
 
 namespace chainerx {
