+++ chainerx_cc/chainerx/native/native_device/misc.cc
@@ -7,6 +7,7 @@
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
 #include "chainerx/kernels/math.h"
+#include "chainerx/kernels/logic.h"
 #include "chainerx/kernels/misc.h"
 #include "chainerx/native/elementwise.h"
 #include "chainerx/native/kernel_regist.h"
