+++ chainerx_cc/chainerx/routines/math.h
@@ -13,12 +13,6 @@ Array Sigmoid(const Array& x);
 
 Array Relu(const Array& x);
 
-Array IsNan(const Array& x);
-
-Array IsInf(const Array& x);
-
-Array IsFinite(const Array& x);
-
 Array Ceil(const Array& x);
 
 Array Floor(const Array& x);
