+++ tests/chainerx_tests/unit_tests/routines_tests/test_math.py
@@ -181,42 +181,3 @@ def test_rounding_routines(func, xp, device, input, dtypes):
     a = func(xp, a)
     a = dtype_utils.cast_if_numpy_array(xp, a, out_dtype)
     return a
-
-
-@chainerx.testing.numpy_chainerx_array_equal()
-@pytest.mark.parametrize_device(['native:0', 'cuda:0'])
-@pytest.mark.parametrize('input', [
-    numpy.asarray(0), numpy.asarray(-1), numpy.asarray(
-        10), numpy.asarray(float('inf')), numpy.asarray(-float('inf')),
-    numpy.asarray(float('nan')), numpy.full(
-        (), 2), numpy.full((0,), 2), numpy.full((2, 3), 2)
-])
-def test_isnan(xp, device, input, dtype):
-    a = xp.array(input.astype(dtype))
-    return xp.isnan(a)
-
-
-@chainerx.testing.numpy_chainerx_array_equal()
-@pytest.mark.parametrize_device(['native:0', 'cuda:0'])
-@pytest.mark.parametrize('input', [
-    numpy.asarray(0), numpy.asarray(-1), numpy.asarray(
-        10), numpy.asarray(float('inf')), numpy.asarray(-float('inf')),
-    numpy.asarray(float('nan')), numpy.full(
-        (), 2), numpy.full((0,), 2), numpy.full((2, 3), 2)
-])
-def test_isinf(xp, device, input, dtype):
-    a = xp.array(input.astype(dtype))
-    return xp.isinf(a)
-
-
-@chainerx.testing.numpy_chainerx_array_equal()
-@pytest.mark.parametrize_device(['native:0', 'cuda:0'])
-@pytest.mark.parametrize('input', [
-    numpy.asarray(0), numpy.asarray(-1), numpy.asarray(
-        10), numpy.asarray(float('inf')), numpy.asarray(-float('inf')),
-    numpy.asarray(float('nan')), numpy.full(
-        (), 2), numpy.full((0,), 2), numpy.full((2, 3), 2)
-])
-def test_isfinite(xp, device, input, dtype):
-    a = xp.array(input.astype(dtype))
-    return xp.isfinite(a)
